+++ internal/controller/patches_test.go
@@ -5,6 +5,7 @@ import (
 
 	cmmetav1 "github.com/cert-manager/cert-manager/pkg/apis/meta/v1"
 	"github.com/stretchr/testify/assert"
+	corev1 "k8s.io/api/core/v1"
 	kusttypes "sigs.k8s.io/kustomize/api/types"
 
 	"github.com/piraeusdatastore/piraeus-operator/v2/internal/controller"
@@ -38,9 +39,71 @@ func TestPatches(t *testing.T) {
 			},
 		},
 		{
+			name: "ClusterLinstorControllerNodeSelector",
+			call: func() ([]kusttypes.Patch, error) {
+				return controller.ClusterLinstorControllerNodeSelector(map[string]string{"foo": "bar"})
+			},
+		},
+		{
+			name: "ClusterLinstorControllerNodeAffinityPatch",
+			call: func() ([]kusttypes.Patch, error) {
+				return controller.ClusterLinstorControllerNodeAffinityPatch(&corev1.NodeSelector{
+					NodeSelectorTerms: []corev1.NodeSelectorTerm{{MatchExpressions: []corev1.NodeSelectorRequirement{{
+						Key:      "example.com/label",
+						Operator: corev1.NodeSelectorOpDoesNotExist,
+					}}}},
+				})
+			},
+		},
+		{
+			name: "ClusterCSIControllerNodeSelector",
+			call: func() ([]kusttypes.Patch, error) {
+				return controller.ClusterCSIControllerNodeSelector(map[string]string{"foo": "bar"})
+			},
+		},
+		{
+			name: "ClusterCSIControllerNodeAffinityPatch",
+			call: func() ([]kusttypes.Patch, error) {
+				return controller.ClusterCSIControllerNodeAffinityPatch(&corev1.NodeSelector{
+					NodeSelectorTerms: []corev1.NodeSelectorTerm{{MatchExpressions: []corev1.NodeSelectorRequirement{{
+						Key:      "example.com/label",
+						Operator: corev1.NodeSelectorOpDoesNotExist,
+					}}}},
+				})
+			},
+		},
+		{
 			name: "ClusterCSINodeSelectorPatch",
 			call: func() ([]kusttypes.Patch, error) {
 				return controller.ClusterCSINodeSelectorPatch(map[string]string{"foo": "bar"})
+			},
+		},
+		{
+			name: "ClusterCSINodeNodeAffinityPatch",
+			call: func() ([]kusttypes.Patch, error) {
+				return controller.ClusterCSINodeNodeAffinityPatch(&corev1.NodeSelector{
+					NodeSelectorTerms: []corev1.NodeSelectorTerm{{MatchExpressions: []corev1.NodeSelectorRequirement{{
+						Key:      "example.com/label",
+						Operator: corev1.NodeSelectorOpDoesNotExist,
+					}}}},
+				})
+			},
+		},
+		{
+			name: "ClusterHAControllerNodeSelectorPatch",
+			call: func() ([]kusttypes.Patch, error) {
+				return controller.ClusterHAControllerNodeSelectorPatch(map[string]string{"foo": "bar"})
+			},
+		},
+		{
+			name: "ClusterHAControllerNodeAffinityPatch",
+			call: func() ([]kusttypes.Patch, error) {
+				return controller.ClusterHAControllerNodeAffinityPatch(&corev1.NodeSelector{
+					NodeSelectorTerms: []corev1.NodeSelectorTerm{{MatchExpressions: []corev1.NodeSelectorRequirement{{
+						Key:      "example.com/label",
+						Operator: corev1.NodeSelectorOpDoesNotExist,
+					}}}},
+				})
 			},
 		},
 		{