+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add image configuration for CSI sidecars.
 - Check kernel module parameters for DRBD on load.
 - Automatically set SELinux labels when loading kernel modules.
+- Allow more complex node selection by adding `LinstorCluster.spec.nodeAffinity`.
 
 ### Changed